bookinventory/requirements.txt
amqp==2.6.1
backcall==0.2.0
billiard==3.6.4.0
celery==4.4.7
decorator==5.1.0
Django==2.2.24
django-filter==2.4.0
djangorestframework==3.12.4
importlib-metadata==2.1.1
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.17.2
kombu==4.6.11
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==2.0.10
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
traitlets==4.3.3
vine==1.3.0
wcwidth==0.2.5
zipp==1.2.0
